Refactored tests, added some test assets
